Move test infrastructure to own directory `test/`
