+++ src/launcher/features/sources/ConfirmRemoveSourceDialog.test.tsx
@@ -6,7 +6,7 @@
 
 import React from 'react';
 
-import render from '../../../testrenderer';
+import render from '../../../test/testrenderer';
 import ConfirmRemoveSourceDialog from './ConfirmRemoveSourceDialog';
 import { showRemoveSource } from './sourcesSlice';
