+++ src/test/README.md
@@ -1,0 +1,2 @@
+This directory contains infrastructure for tests. The concrete tests are placed
+next to the units under tests.
